cleanup commit

+++ src/main/java/budgeter/Transaction.java
@@ -1,18 +1,20 @@
 package budgeter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
+
 
 public class Transaction {
     private int transactionID;
     private LocalDate date;
     private LocalDate weekStartDate;
-    private Double total;
+    private BigDecimal total;
     private CategoryId categoryID;
     private String paymentType;
     private String title;
     private String vendor;
 
-    public Transaction(int year, int month, int day, Double totalArg, String paymentTypeArg, String vendorArg) {
+    public Transaction(int year, int month, int day, BigDecimal totalArg, String paymentTypeArg, String vendorArg) {
         date = LocalDate.of(year, month, day);
         weekStartDate = date.minusDays(date.getDayOfWeek().getValue() - 1);
         total = totalArg;
@@ -40,7 +42,8 @@ public class Transaction {
         return date;
     }
 
-    public Double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
+
 }
